refactored export_to_repository_ckan
Missing documentation for a few functions in ckanUtils
